requirements.txt
asyncpg==0.30.0
fastapi==0.115.12
greenlet==3.2.2
pydantic==2.11.5
python-dotenv==1.1.0
SQLAlchemy==2.0.41
uvicorn==0.34.2
